Module type Parameters.GENERIC_ELT

A type for generic elements stored in our Functional union-find structures. This should be a super-type of PatriciaTree.HETEROGENEOUS_KEY.

include SIMPLE_GENERIC_ELT

type 'a t

The type of elements (nodes) in the union-find structure

val polyeq : 'a t -> 'b t -> ('a, 'b) PatriciaTree.cmp

polymorphic equality on elements

val to_int : 'a t -> int

Returns a unique integer identifier associated with the given element (eg hash-consed tag). See PatriciaTree.HETEROGENEOUS_KEY.to_int

val pretty : Format.formatter -> 'a t -> unit

Pretty printer